Update Signup flow and texts.
This should accomplish most of what we've discussed in several
meetngs in Stockholm this week.

E-mail confirmation is now required before being given the option
to proceed to the dashboard.

E-mails with passwords are disabled by default now (I think).

Signup flow should be easier to follow now, and everything is
prepared for accepting a finish_url from the SP, although we
currently just get a default URL from the config.

# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.