More tests and bug fixes (one important).
Important bug fix:

Don't care about unverified e-mail addresses in the central eduid
userdb.
Signup used to send out new confirmation codes when someone signed up
with an unconfirmed e-mail address, and if that configuration code was
used someone might end up with two accounts with the same e-mail address
confirmed (requires access to the e-mail account, so it is not a way to
steal e-mail addresses).

This could render one or both the accounts inaccessible, as the IdP
verifies that one and only one account is returned when searching for
users using the e-mail address.